+++ entity/src/record.rs
@@ -8,13 +8,13 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "records")]
 pub struct Model {
     #[sea_orm(primary_key)]
-    pub id: i64,
-    pub user_id: i64,
+    pub id: i32,
+    pub user_id: i32,
 
     #[sea_orm(column_type = "Binary(BlobSize::Blob(None))")]
     pub user_pubkey: Vec<u8>,
 
-    pub invoice_id: Option<i64>,
+    pub invoice_id: Option<i32>,
 
     /// user current balance in msats
     pub balance: i64,
